==> 1_installation BD EUTL.ipynb <==
voir toutes les indications https://github.com/jabrell/pyeutl/tree/main

Code à lancer:
`i, n 
pip install git+https://github.com/jabrell/pyeutl.git`

Installer postgresSQL version 16

import pandas as pd

2009 année 0
banking 2009 = quotas alloués gratuitement - émissions vérifiées/ quotas restitués + quotas acquis aux enchères (marché primaire) + montants nets des actions sur le marchés secondaire 

Grouper par secteur

Test 1 secteur du ciment, id = 23.51

from pyeutl.orm import DataAccessLayer

connectionSettings = dict(
    user="postgres", 
    host="localhost", 
    db="eutl2024", 
    passw="password",
    port=5432
)



- Rajouter dans PATH des variables système le lien vers POSTGRES:
`C:\Program Files\PostgreSQL\17\bin`
- Puis lancer cette commande dans le terminal pour s'assurer du bon fonctionnement des credentials donnés ci-dessus et de POSTGRES:
`psql -h localhost -p 5432 -U postgres -W`

- Si le code d'en dessous ne fonctionne toujours pas, créer une base de données appelée EU ETL
en relançant:
``psql -h localhost -p 5432 -U postgres -W``

puis en entrant le mdp: password et puis en compilant:
postgres=#`CREATE DATABASE eutl2024`;

-> j'ai appuyé sur echap lors du pop up qui demande si il peut supprimer les tables existantes (mais peut etre inutile vue qu'il n'y a pas de tables dans eutl2024)



#création de la base de données
dal = DataAccessLayer(**connectionSettings)
#creation de la base


dal.create_database()
