==> 3_stationnarité.ipynb <==
CLUSTER 0


import matplotlib.pyplot as plt
import matplotlib.dates as mdates
import pandas as pd

phases = {
    'Phase II': ['2008-01-01', '2012-12-01'],
    'Phase III': ['2013-01-01', '2020-12-01']
}



Data=pd.read_csv("combined_data/cluster0.csv")
Data= Data.drop(columns=["Unnamed: 0","Time","month"])
Data= Data.drop(columns=["Unnamed: 0.1"])



Data['Date']=pd.to_datetime(Data['Date'])

# Créer des DataFrames séparés pour chaque phase
phase_data = {}
for phase, (start, end) in phases.items():
    start_date = pd.to_datetime(start)
    end_date = pd.to_datetime(end)
    phase_data[phase] = Data[(Data['Date'] >= start_date) & (Data['Date'] <= end_date)]

# Afficher les premières lignes de chaque DataFrame de phase
for phase, df in phase_data.items():
    # Utiliser exec pour créer dynamiquement des variables
    exec(f"{phase.replace(' ', '_')}_df = df")

from statsmodels.tsa.stattools import adfuller

def test_stationarity(series, significance_level=0.05):
    result = adfuller(series.dropna())
    p_value = result[1]
    return p_value   # True si stationnaire, False sinon


import numpy as np

def make_sationary(data):
    # Calcul du décalage nécessaire pour rendre toutes les valeurs positives avant de prendre le logarithme
    shift_per_col = data.drop(columns=['Date']).min().abs() + 1
    data_stat = data.copy()

    # Appliquer une transformation logarithmique et des différenciations d'ordre 1 et 2 si nécessaire
    for col in data_stat.columns:
        if col != 'Date':
            data_stat[col] += shift_per_col[col]
            data_stat[col] = np.log1p(data_stat[col])
            p_value_after_log = test_stationarity(data_stat[col])  # Tester la stationnarité

            if p_value_after_log < 0.05:
                print(f"✅ Stationnaire après transformation log pour {col}.")
            else:
                # Appliquer une différenciation d'ordre 1
                data_stat[col] = data_stat[col].diff()
                p_value_after_diff1 = test_stationarity(data_stat[col])

                if p_value_after_diff1 < 0.05:
                    print(f"📉✅ Différenciation d'ordre 1 appliquée à {col} et p-value après: {p_value_after_diff1}")
                else:
                    # Appliquer une différenciation d'ordre 2
                    data_stat[col] = data_stat[col].diff()
                    p_value_after_diff2 = test_stationarity(data_stat[col])
                    print(f"📉📉 Différenciation d'ordre 2 appliquée à {col} et est stationnaire: {(p_value_after_diff2<0.05)}")
    return data_stat


new_phase_II_df = make_sationary(Phase_II_df)


for col in new_phase_II_df.drop(columns=['Date']).columns:
    p_value = test_stationarity(new_phase_II_df[col])
    if p_value < 0.05:
        print(f"La série {col} est stationnaire")
    else:
        print(f"La série {col} n'est pas stationnaire (p-value >= {0.05}).")
new_phase_II_df.to_csv("data_stationnaires/Cluster0_Phase_II.csv")

new_phase_III_df = make_sationary(Phase_III_df)

for col in new_phase_III_df.drop(columns=['Date']).columns:
    p_value = test_stationarity(new_phase_III_df[col])
    if p_value < 0.05:
        print(f"La série {col} est stationnaire")
    else:
        print(f"La série {col} n'est pas stationnaire (p-value >= {0.05}).")
new_phase_III_df.to_csv("data_stationnaires/Cluster0_Phase_III.csv")

CLUSTER 1

Data1=pd.read_csv("combined_data/cluster1.csv")
Data1= Data1.drop(columns=["Unnamed: 0","Time","month"])
Data1= Data1.drop(columns=["Unnamed: 0.1"])
Data1.head()

Data1['Date']=pd.to_datetime(Data1['Date'])
# Créer des DataFrames séparés pour chaque phase
phase_data = {}
for phase, (start, end) in phases.items():
    start_date = pd.to_datetime(start)
    end_date = pd.to_datetime(end)
    phase_data[phase] = Data1[(Data1['Date'] >= start_date) & (Data1['Date'] <= end_date)]

# Afficher les premières lignes de chaque DataFrame de phase
for phase, df in phase_data.items():
    # Utiliser exec pour créer dynamiquement des variables
    exec(f"{phase.replace(' ', '_')}_df = df")

newcluster1_phase_II_df = make_sationary(Phase_II_df)
for col in newcluster1_phase_II_df.drop(columns=['Date']).columns:
    p_value = test_stationarity(newcluster1_phase_II_df[col])
    if p_value < 0.05:
        print(f"La série {col} est stationnaire")
    else:
        print(f"La série {col} n'est pas stationnaire (p-value >= {0.05}).")
newcluster1_phase_II_df.to_csv("data_stationnaires/Cluster1_Phase_II.csv")

newcluster1_phase_III_df = make_sationary(Phase_III_df)
for col in newcluster1_phase_III_df.drop(columns=['Date']).columns:
    p_value = test_stationarity(newcluster1_phase_III_df[col])
    if p_value < 0.05:
        print(f"La série {col} est stationnaire")
    else:
        print(f"La série {col} n'est pas stationnaire (p-value >= {0.05}).")
newcluster1_phase_III_df.to_csv("data_stationnaires/Cluster1_Phase_III.csv")
