==> 4_lags-length.ipynb <==
import pandas as pd

cluster1_phase2 = pd.read_csv('data_stationnaires/Cluster1_Phase_II.csv')
cluster1_phase3 = pd.read_csv('data_stationnaires/Cluster1_Phase_III.csv')
cluster0_phase2 = pd.read_csv('data_stationnaires/Cluster0_Phase_II.csv')
cluster0_phase3 = pd.read_csv('data_stationnaires/Cluster0_Phase_III.csv')


len(cluster1_phase3.columns)

cluster1_phase2.drop(columns=['Unnamed: 0'],inplace=True)
cluster1_phase3.drop(columns=['Unnamed: 0'],inplace=True)
cluster0_phase3.drop(columns=['Unnamed: 0'],inplace=True)
cluster0_phase2.drop(columns=['Unnamed: 0'],inplace=True)
cluster1_phase2.dropna(inplace=True)
cluster1_phase3.dropna(inplace=True)
cluster0_phase2.dropna(inplace=True)
cluster0_phase3.dropna(inplace=True)

import numpy as np
import statsmodels.api as sm

from statsmodels.tsa.ar_model import AutoReg




# Déterminer le nombre optimal de lags
def determine_lags(data):
    AIC = {}
    BIC = {}
    HQ = {}
    T = len(data)
    maxlags = int(4 * (T / 100)**(2/9)) 
    for lag in range(0, maxlags+1):  
        model = AutoReg(data, lags=lag, old_names=False)
        model_fitted = model.fit()
        AIC[lag] = model_fitted.aic
        BIC[lag] = model_fitted.bic
        HQ[lag] = model_fitted.hqic
        
    print(AIC)
    minAIC = min(AIC, key=AIC.get)
    minBIC = min(BIC, key=BIC.get)
    minHQ = min(HQ, key=HQ.get)
    
    return minAIC, minBIC, minHQ

# Appliquer la fonction à vos données




data = cluster1_phase2.copy()
cluster1_2_lags = dict()
for c in data.columns:
    if c != 'Date':
        min_aic_lag, min_bic_lag,min_hq_lag = determine_lags(data[c])
        cluster1_2_lags[c]=(min_aic_lag+ min_bic_lag+min_hq_lag)//3

np.average(list(cluster1_2_lags.values()))

data = cluster0_phase2.copy()
cluster0_2_lags = dict()
for c in data.columns:
    if c != 'Date':
        min_aic_lag, min_bic_lag,min_hq_lag = determine_lags(data[c])
        cluster0_2_lags[c]=(min_aic_lag+ min_bic_lag+min_hq_lag)//3
        
data = cluster0_phase3.copy()
cluster0_3_lags = dict()
for c in data.columns:
    if c != 'Date':
        min_aic_lag, min_bic_lag,min_hq_lag = determine_lags(data[c])
        cluster0_3_lags[c]=(min_aic_lag+ min_bic_lag+min_hq_lag)//3

data = cluster1_phase3.copy()
cluster1_3_lags = dict()
for c in data.columns:
    if c != 'Date':
        min_aic_lag, min_bic_lag,min_hq_lag = determine_lags(data[c])
        cluster1_3_lags[c]=(min_aic_lag+ min_bic_lag+min_hq_lag)//3


print(np.average(list(cluster1_2_lags.values())))
print(np.average(list(cluster1_3_lags.values())))
print(np.average(list(cluster0_2_lags.values())))
print(np.average(list(cluster0_3_lags.values())))

Ajout de 2 lags par variable

Cluster 0 Phase 2

lag_length = 2
data = cluster0_phase2
# Liste pour stocker les colonnes avec les lags
lagged_data = []

# Ajouter des lags sur les colonnes sélectionnées
for col in data.columns:
    if col != 'Date' :
        # Ajouter des lags pour la colonne, avec shift
        for lag in range(1, lag_length + 1):  # Ajouter plusieurs lags
            lagged_col = data[col].shift(lag)
            lagged_col.name = f"{col}_lag{lag}"  # Nommer la colonne laggée
            lagged_data.append(lagged_col)

# Concaténer les colonnes originales avec les nouvelles colonnes avec lags
data = pd.concat([data] + lagged_data, axis=1)

# Afficher le DataFrame avec les lags


data.to_csv('data_stationnaires/Cluster0_Phase_II.csv')

lag_length = 2
data = cluster0_phase3
# Liste pour stocker les colonnes avec les lags
lagged_data = []

# Ajouter des lags sur les colonnes sélectionnées
for col in data.columns:
    if col != 'Date' :
        # Ajouter des lags pour la colonne, avec shift
        for lag in range(1, lag_length + 1):  # Ajouter plusieurs lags
            lagged_col = data[col].shift(lag)
            lagged_col.name = f"{col}_lag{lag}"  # Nommer la colonne laggée
            lagged_data.append(lagged_col)

# Concaténer les colonnes originales avec les nouvelles colonnes avec lags
data = pd.concat([data] + lagged_data, axis=1)

# Afficher le DataFrame avec les lags


data.to_csv('data_stationnaires/Cluster0_Phase_III.csv')

lag_length = 2
data = cluster1_phase2
# Liste pour stocker les colonnes avec les lags
lagged_data = []

# Ajouter des lags sur les colonnes sélectionnées
for col in data.columns:
    if col != 'Date' :
        # Ajouter des lags pour la colonne, avec shift
        for lag in range(1, lag_length + 1):  # Ajouter plusieurs lags
            lagged_col = data[col].shift(lag)
            lagged_col.name = f"{col}_lag{lag}"  # Nommer la colonne laggée
            lagged_data.append(lagged_col)

# Concaténer les colonnes originales avec les nouvelles colonnes avec lags
data = pd.concat([data] + lagged_data, axis=1)

# Afficher le DataFrame avec les lags


data.to_csv('data_stationnaires/Cluster1_Phase_II.csv')

lag_length = 2
data = cluster1_phase3
# Liste pour stocker les colonnes avec les lags
lagged_data = []

# Ajouter des lags sur les colonnes sélectionnées
for col in data.columns:
    if col != 'Date' :
        # Ajouter des lags pour la colonne, avec shift
        for lag in range(1, lag_length + 1):  # Ajouter plusieurs lags
            lagged_col = data[col].shift(lag)
            lagged_col.name = f"{col}_lag{lag}"  # Nommer la colonne laggée
            lagged_data.append(lagged_col)

# Concaténer les colonnes originales avec les nouvelles colonnes avec lags
data = pd.concat([data] + lagged_data, axis=1)

# Afficher le DataFrame avec les lags


data.to_csv('data_stationnaires/Cluster1_Phase_III.csv')
