==> 6_creation_periodes.ipynb <==
import pandas as pd

cluster0_2 = pd.read_csv('data_stationnaires/Cluster0_Phase_II.csv')
cluster0_3 = pd.read_csv('data_stationnaires/Cluster0_Phase_III.csv')
cluster1_2 = pd.read_csv('data_stationnaires/Cluster1_Phase_II.csv')
cluster1_3 = pd.read_csv('data_stationnaires/Cluster1_Phase_III.csv')


cluster0 = pd.concat([cluster0_2,cluster0_3])
cluster1 = pd.concat([cluster1_2,cluster1_3])

cluster0.sort_values(by='Date',inplace=True)
cluster1.sort_values(by='Date',inplace=True)

cluster0.drop(columns=['Unnamed: 0'],inplace=True)
cluster1.drop(columns=['Unnamed: 0'],inplace=True)

breakpoint = '2011-12-01'
cluster0_periode1 = cluster0[cluster0['Date'] < breakpoint].dropna().sort_values(by='Date')
cluster0_periode2 = cluster0[cluster0['Date'] >= breakpoint].dropna().sort_values(by='Date')

cluster1_periode1 = cluster1[cluster1['Date'] < breakpoint].dropna().sort_values(by='Date')
cluster1_periode2 = cluster1[cluster1['Date'] >= breakpoint].dropna().sort_values(by='Date')

cluster0_periode1.to_csv('data_stationnaires/Cluster0_periode1.csv')
cluster0_periode2.to_csv('data_stationnaires/Cluster0_periode2.csv')

cluster1_periode1.to_csv('data_stationnaires/Cluster1_periode1.csv')
cluster1_periode2.to_csv('data_stationnaires/Cluster1_periode2.csv')
