==> 7_multicolinearite.ipynb <==
import pandas as pd
import warnings
from sklearn.exceptions import ConvergenceWarning
warnings.filterwarnings("ignore", category=ConvergenceWarning)


cluster1_phase2 = pd.read_csv('data_stationnaires/Cluster1_Phase_II.csv')
cluster1_phase3 = pd.read_csv('data_stationnaires/Cluster1_Phase_III.csv')
cluster0_phase2 = pd.read_csv('data_stationnaires/Cluster0_Phase_II.csv')
cluster0_phase3 = pd.read_csv('data_stationnaires/Cluster0_Phase_III.csv')
cluster1_periode1 = pd.read_csv('data_stationnaires/Cluster1_periode1.csv')
cluster1_periode2 = pd.read_csv('data_stationnaires/Cluster1_periode2.csv')
cluster0_periode1 = pd.read_csv('data_stationnaires/Cluster0_periode1.csv')
cluster0_periode2 = pd.read_csv('data_stationnaires/Cluster0_periode2.csv')

from statsmodels.stats.outliers_influence import variance_inflation_factor
import statsmodels.api as sm


def calculate_vifs(dataframe):
    vifs = pd.DataFrame()
    vifs["Variable"] = dataframe.columns
    vifs["VIF"] = [variance_inflation_factor(dataframe.values, i) for i in range(dataframe.shape[1])]
    return vifs

def clean_data(X):
    removed_vars = []

    # Boucle jusqu'à ce que toutes les variables aient un VIF inférieur à 10
    while True:
        VIFs = calculate_vifs(X)
        # Trouve la variable avec le VIF maximum qui n'est pas const
        max_vif = VIFs[VIFs["Variable"] != 'const'].max()
        if max_vif["VIF"] < 10:
            break  # Sort de la boucle si toutes les variables ont un VIF < 10
        variable_to_remove = VIFs[VIFs["VIF"] == max_vif["VIF"]]["Variable"].values[0]
        X.drop(columns=[variable_to_remove], inplace=True)
        removed_vars.append(variable_to_remove)
        
    print("Variables removed:", removed_vars)
    print(VIFs)
    return removed_vars



X = cluster1_phase3.drop(columns=['Price', 'Date', 'Unnamed: 0'])
X = sm.add_constant(X)
X.dropna(inplace=True)
removed_vars = clean_data(X)
cluster1_phase3.drop(columns=removed_vars, inplace=True) 
cluster1_phase3.shape


X = cluster1_phase2.drop(columns=['Price', 'Date', 'Unnamed: 0'])
X = sm.add_constant(X)
X.dropna(inplace=True)
removed_vars = clean_data(X)
cluster1_phase2.drop(columns=removed_vars, inplace=True) 
cluster1_phase2.shape

X = cluster0_phase3.drop(columns=['Price', 'Date', 'Unnamed: 0'])
X = sm.add_constant(X)
X.dropna(inplace=True)
removed_vars = clean_data(X)
cluster0_phase3.drop(columns=removed_vars, inplace=True) 
cluster0_phase3.shape

X = cluster0_phase2.drop(columns=['Price', 'Date', 'Unnamed: 0'])
X = sm.add_constant(X)
X.dropna(inplace=True)
removed_vars = clean_data(X)
cluster0_phase2.drop(columns=removed_vars, inplace=True) 
cluster0_phase2.shape

X = cluster0_periode1.drop(columns=['Price', 'Date', 'Unnamed: 0'])
X = sm.add_constant(X)
X.dropna(inplace=True)
removed_vars = clean_data(X)
cluster0_periode1.drop(columns=removed_vars, inplace=True) 
cluster0_periode1.shape

X = cluster0_periode2.drop(columns=['Price', 'Date', 'Unnamed: 0'])
X = sm.add_constant(X)
X.dropna(inplace=True)
removed_vars = clean_data(X)
cluster0_periode2.drop(columns=removed_vars, inplace=True) 
cluster0_periode2.shape

X = cluster1_periode1.drop(columns=['Price', 'Date', 'Unnamed: 0'])
X = sm.add_constant(X)
X.dropna(inplace=True)
removed_vars = clean_data(X)
cluster1_periode1.drop(columns=removed_vars, inplace=True) 
cluster1_periode1.shape

X = cluster1_periode2.drop(columns=['Price', 'Date', 'Unnamed: 0'])
X = sm.add_constant(X)
X.dropna(inplace=True)
removed_vars = clean_data(X)
cluster1_periode2.drop(columns=removed_vars, inplace=True) 
cluster1_periode2.shape

cluster1_phase2.to_csv('data_stationnaires/Cluster1_Phase_II.csv')
cluster1_phase3.to_csv('data_stationnaires/Cluster1_Phase_III.csv')
cluster0_phase2.to_csv('data_stationnaires/Cluster0_Phase_II.csv')
cluster0_phase3.to_csv('data_stationnaires/Cluster0_Phase_III.csv')
cluster1_periode1.to_csv('data_stationnaires/Cluster1_periode1.csv')
cluster1_periode2.to_csv('data_stationnaires/Cluster1_periode2.csv')
cluster0_periode1.to_csv('data_stationnaires/Cluster0_periode1.csv')
cluster0_periode2.to_csv('data_stationnaires/Cluster0_periode2.csv')
